Update base_connection.py
If we want to provide read_timeout same for most of the commands and only few commands have different read_timeout then we need to either make the read_timeout_override maximum accroding to the command set or remove this attribute. But after this change we will be able to read the command line in read_timeout by bypassing read_timeout_override  time (for send_command function only).